requirements.txt
pyyaml
irc
ib3
mwclient
